main: improve version string
Initialize a build variable to the date on which the binary was
generated. Include the date in the build info string which is output
when running aerc -v and in the crash logs.

Do not rely on parsing the build flags via some obscure base64 voodoo to
determine if notmuch support is available or not. Instead, use the
symbols from the linked library directly if available.

Before:

 $ aerc -v
 0.16.0-183-g4cc2e6be3a01 +notmuch (go1.21.6 amd64 linux)

After:

 $ aerc -v
 aerc 0.16.0-183-g4cc2e6be3a01 +notmuch-5.6.0 (go1.21.6 amd64 linux 2024-01-31)

Signed-off-by: Robin Jarry <robin@jarry.cc>
Acked-by: Ciarán Ainsworth <cda@sporiff.dev>