filter: allow workers to combine filter terms
Allow the backend workers to combine the filter terms. Currently, the
consecutive filters are joined in the message store with a space (" ").
This works well for most backends, but makes the filter combination for
notmuch confusing.

Example:

Issuing two consecutive filter commands in notmuch

	:filter not tag:list
	:filter tag:list

would create the following filter query

	'not tag:list tag:list'

This is not what users would expect; they expect:

	'(not tag:list) and (tag:list)'

Note that the notmuch backend works correctly for the given query, but
produced a query that does not match the user's expectation. This patch
fixes this.

The combination of filter terms in other backends remains the same.

Reported-by: Ángel Castañeda <angel@acsq.me>
Signed-off-by: Koni Marti <koni.marti@gmail.com>
Tested-by: Inwit <inwit@sindominio.net>
Acked-by: Robin Jarry <robin@jarry.cc>