config: change default appearance
The default styleset and ui settings are quite dull and don't make use
of the "newer" features that have been added to aerc in the past two
years.

Change the default styleset to something more colorful that renders
correctly on both dark and light backgrounds. Change the default ui
settings to get something more appealing to new users.

Save the previous default color theme to a new `monochrome` styleset.

Changelog-changed: The default `[ui]` settings and the `default`
 styleset have changed extensively. A no-color theme can be restored
 with the `monochrome` styleset.
Cc: Marc Coquand <marc@mccd.space>
Signed-off-by: Robin Jarry <robin@jarry.cc>
Tested-by: Nicolas Dichtel <nicolas.dichtel@6wind.com>