templates: use email address as fallback
If replying to mail or forwarding a mail that doesn't have a
recipient name, the automatically generated header line will just
have an empty value. Use the names template to extract either the
name if available otherwise the mbox name.

While at it, instead of just taking the first item of the list,
join all items to avoid an index error if the list is empty.

Signed-off-by: Michael Walle <michael@walle.cc>
Acked-by: Robin Jarry <robin@jarry.cc>