binds: allow account and folder binds to coexist
When loading binds, account-specific binds are first applied, and then
folder-specific binds are applied. However, in the process of merging
the account-specific binds with the general binds, the folder-specific
binds are lost. As a result, if both account- and folder-specific binds
exist, the folder-specific binds will do nothing when in the specified
account. The following minimal binds.conf reproduces this behavior:

    [messages]
    [messages:account=<ACCOUNT>]
    [messages:folder=<FOLDER>]
    q = :quit<Enter>

Fix this issue by ensuring contextual binds are preserved when merging.

Changelog-fixed: Allow account- and folder-specific binds to coexist.
Signed-off-by: Jason Cox <me@jasoncarloscox.com>
Acked-by: Robin Jarry <robin@jarry.cc>