msglist: allow configuring default split
Add a new setting to configure the default split layout for message list
tabs. The syntax is a bit different from the :split and :vsplit commands
since it needs to convey the direction in the value as well. I didn't
reuse split/vsplit since they are a bit confusing when used in
a configuration file.

The syntax is as follows:

	message-list-split = [<direction>] <size>

The direction is optional and defaults to horizontal. The size is the
number of terminal cells that will be used to display the message list.

All these examples are equivalent:

	message-list-split = horiz 12
	message-list-split = h 12
	message-list-split = 12

Same idea for vertical splits:

	message-list-split = vertical 120
	message-list-split = vert 120
	message-list-split = v 120

Both :split and :vsplit commands remain usable as before. The
configuration options only affect the initial layout at startup.

Add config.SPLIT_* constants and sanitize AccountView.{Split,Vsplit}
methods.

Changelog-added: Configure default message list `:split` or `:vsplit`
 on startup with `message-list-split` in `aerc.conf`.
Signed-off-by: Robin Jarry <robin@jarry.cc>
Tested-by: Tim Culverhouse <tim@timculverhouse.com>