mod: update all dependencies
Patch produced with the following commands:

	go mod edit -dropreplace=golang.org/x/crypto
	go get -t -u
	go mod tidy

I dropped the replace of x/crypto. The only use for the fork from
ProtonMail is for openpgp which has been deprecated by the official
implementation. And aerc explicitly uses the correct import path where
needed.

go-smtp update was a bit harsh. The API was changed completely.

Signed-off-by: Robin Jarry <robin@jarry.cc>
Tested-by: Bence Ferdinandy <bence@ferdinandy.com>