send: prevent sending if multipart conversion failed
Given the following configuration:

binds.conf:
	[compose::review]
	y = :multipart text/html<Enter>:send<Enter>

aerc.conf:
	[multipart-converters]
	text/html = /path/to/some/script.sh

/path/to/some/script.sh:
	#!/bin/sh
	exit 10 # falls for some reason

When you press `y` aerc runs `:multipart` command and although it gets
an error from the converter script, the error is ignored and then the
`:send` command actually sends a broken message.

Add ConversionError field to Composer.Part to track multipart conversion
errors.

Check for conversion errors in :send, block sending if the errors are
found.

There is no way to skip this like missing attachment or empty subject.
This is done intentionally. The user needs to update or delete the
problematic part before actually sending a message.

Signed-off-by: Vitaly Ovchinnikov <v@ovch.ru>
Acked-by: Robin Jarry <robin@jarry.cc>