app: make aerc satisfy the Beeper interface again
During the Vaxis switchover, the Beeper interface was changed because
Vaxis' Bell() function does not return error. However the corresponding
change was not made on aerc making it not satisfy this interface and
thus the DrawableInteractiveBeeper interface. This lead to the Bell
function not getting registered making aerc unable to ring the terminal
bell. Fix this.

Fixes: 0fd5f4115792 ("ui: initialize vaxis directly, drop tcell.Screen initialization")
Signed-off-by: Karel Balej <balejk@matfyz.cz>
Reviewed-by: Tim Culverhouse <tim@timculverhouse.com>
Acked-by: Robin Jarry <robin@jarry.cc>