doc: update and rephrase -H header search syntax
The syntax for the -H option's argument was confusing, warning about
requiring the value to be from the header by a space without hinting
that it also expect a single argument: requiring the query to be wrapped
with quotes.

Since 8464b3738 ("search: use a common api for all workers") the parser
and the syntax has changed, now: the query is split around `:` without
needing a space.

Update the man page to reflect this change.

Signed-off-by: Jules Maselbas <jmaselbas@zdiv.net>
Reviewed-by: Tim Culverhouse <tim@timculverhouse.com>
Acked-by: Robin Jarry <robin@jarry.cc>