accounts: skip accounts with errors instead of exiting
Currently, if something is wrong with an account, say the password
command has issues, aerc will exit. When you have multiple accounts, not
starting any of them does not make much sense. Instead of exiting, drop
the account that failed to parse and log the appropriate error messages.

Since the current behaviour when starting aerc with the -a flag is that
if the user specifies an account that doesn't exist, aerc exits with an
error, the above modification will still lead to aerc exiting when using
the -a flag and having an unparsable account. Instead of exiting, log an
error message.

The only sideeffect of the above modifications, is that in case all the
potential accounts are dropped, the user will be shown the new account
dialog instead of exiting.

Changelog-changed: Unparsable accounts are skipped, instead of aerc
 exiting with an error.
Signed-off-by: Bence Ferdinandy <bence@ferdinandy.com>
Acked-by: Robin Jarry <robin@jarry.cc>