next,prev: fix panic when message list is empty
When running :filter -u and iterating over all messages with maildir,
the list will eventually become empty (e.g. all messages have been
read, no unread messages can be displayed).

At this point, :next or :prev will fail to select anything and the
maildir backend will return an error. Since the error is not caught, the
message viewer is reloaded with a broken message. At this point, any key
press causes a panic.

This only happens with maildir because message list filters are updated
dynamically in the background everytime a filsystem event is triggered.

Avoid reloading the message viewer if an error is returned by the
backend. Just display the error and abort.

Link: https://lists.sr.ht/~rjarry/aerc-devel/%3CD1761AO8LJUL.12698V9VQDR9B@gmail.com%3E
Reported-by: Matthew Bystrin <dev.mbstr@gmail.com>
Signed-off-by: Robin Jarry <robin@jarry.cc>
Tested-by: Koni Marti <koni.marti@gmail.com>