main: detect invalid arguments early
Currently, when running aerc --help or aerc --version, we get an obscure
error when aerc is already running.

	$ aerc --help
	response: command not understood

When it is not running, it starts but prints the same error message in
the status line with weird colors.

Avoid sending garbage into the control socket. Try to detect invalid
arguments early.

	$ aerc --foobaz
	error: unknown argument: --foobaz

Reported-by: Peter Sanchez <pjs@petersanchez.com>
Signed-off-by: Robin Jarry <robin@jarry.cc>
Tested-by: Bence Ferdinandy <bence@ferdinandy.com>