mk: handle stale files when scdoc fails
If scdoc fails for any reason, for example because a user hasn't
installed it yet, the `> $$@` redirection will still create an empty
destination file. On re-runs gmake will see the empty file and assume
the target has already completed. This led to installing a blank
man page.

Use the GNU Make '.DELETE_ON_ERROR' special target which automatically
deletes the target if a recipe fails.

Link: https://www.gnu.org/software/make/manual/html_node/Special-Targets.html#index-removing-targets-on-failure
Signed-off-by: Brandon Maier <brandon.maier@collins.com>
Acked-by: Robin Jarry <robin@jarry.cc>