log: create directories if necessary
If a log file is configured via `log-file` and the parent directory does
not exist then aerc exits with an error message on startup.

Check if the directory for the log file exists and create it if
necessary using sensible permissions (0700). Errors occurring while
creating the directory structure are returned and printed out.

Changelog-fixed: Startup error if `log-file` directory does not exist.
Signed-off-by: Jens Grassel <jens@wegtam.com>
Reviewed-by: Moritz Poldrack <moritz@poldrack.dev>
Acked-by: Robin Jarry <robin@jarry.cc>