calendar: hide empty attendee list
Include the ATTENDEE and DETAILED LIST section in the calendar filter's
output only if there are actually attendees to list.  I was especially
confused by the empty DETAILED LIST, as it wasn't clear which kind of
detail it was supposed to list (and it also looked a bit like a heading
for the description following immediately below), see e.g. the test
output in filters/vectors/calendar-invite.expected.

Fixes: 777bbb77e806 ("contrib: improve readability of meeting requests")
Changelog-fixed: Builtin `calendar` filter shows empty attendee list.
Signed-off-by: Timon Reinold <tirei+aerc@agon.one>
Acked-by: Robin Jarry <robin@jarry.cc>