notmuch: speed up lastmod query.
When changing the tag/label of a message we refresh all messages that
were modified since the start of the transaction.
To that end we add "lastmod before..now" to the current query without
any conjunction.  That would make us re-index basically all messages.

This patch wraps the existing query in parentheses and
ANDs it with "lastmod" instead which greatly speeds up the process.

Signed-off-by: Robin Dapp <rdapp@modk.org>
Acked-by: Robin Jarry <robin@jarry.cc>