aerc: add quake-mode terminal
Add a drop-down (quake-mode) terminal which is a persistent terminal
session that overlays aerc at the top and can be toggled on or off.

Enable quake mode by setting [General].enable-quake-mode=true.

The height of the drop-down terminal can be set with
[ui].quake-terminal-height (default: 20).

Toggling is hardcoded to the F1 key. Note that this key should not be
used in your key bindings when you enable Quake mode.

Signed-off-by: Koni Marti <koni.marti@gmail.com>
Tested-by: Inwit <inwit@sindominio.net>
Acked-by: Robin Jarry <robin@jarry.cc>