patch: add auto-switch option
Add an auto-switch option that changes the project of the patch manager
based on the subject line of a message if it contains a '[PATCH
<project>]' segment.

A subject line with '[PATCH aerc v2]' would switch to the 'aerc' project
if that project is available in the patch manager.

The auto switching can be activated per account by adding
'pama-auto-switch = true' to your account config.

Implements: https://todo.sr.ht/~rjarry/aerc/226
Changelog-added: Auto-switch projects based on the message subject
 for the :patch command.
Signed-off-by: Koni Marti <koni.marti@gmail.com>
Acked-by: Robin Jarry <robin@jarry.cc>