gpg: fix signed message encoding
Fix the content encoding for GPG-signed messages.

To remove the Mime-Version header field for the signed message part, the
raw message is parsed with go-message. go-message.Read(), however,
decodes the message body as well (i.e. from quoted-printable to UTF8
depending on the Content-Transfer-Encoding header). This means that the
msg.Body field now contains the decoded message (it is no longer encoded
as quoted-printable). We never encode the message back to the proper
Content-Transfer-Encoding.

To fix this, use net/mail.ReadMessage() to parse the headers and to not
decode the message body.

To verify the issue, send a signed message with the following text:
"19+1=20!"

The message will be properly signed, but the text is wrong; it shows
"19+1 !"; instead it should read "19+1=3D20!".

Fixes: 5e443bce ("gpg: fix mime-version header position")
References: https://todo.sr.ht/~rjarry/aerc/79
Signed-off-by: Koni Marti <koni.marti@gmail.com>
Tested-by: Jens Grassel <jens@wegtam.com>
Acked-by: Robin Jarry <robin@jarry.cc>